requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
atproto==0.0.65
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
braceexpand==0.1.7
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
discord.py==2.6.4
filelock==3.20.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
gradio_client==2.0.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
imagehash==4.3.1
import_expression==2.2.1.post1
jishaku==2.6.3
Js2Py_3.13==0.74.1
lxml==6.0.2
m3u8==6.0.0
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pyjsparser==2.7.1
pyotp==2.9.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
six==1.17.0
socksio==1.0.0
soupsieve==2.8.1
tabulate==0.9.0
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
webvtt-py==0.5.1
yarl==1.22.0
